bugfix(Inventory): Move inventory to demo package
- fixing architecture diagram, I moved `InventoryObject` and its render helper to the demo package.
