+++ src/main/java/fr/snapgames/fromclasstogame/demo/render/InventoryRenderHelper.java
@@ -1,7 +1,8 @@
-package fr.snapgames.fromclasstogame.core.gfx.renderer;
+package fr.snapgames.fromclasstogame.demo.render;
 
 import fr.snapgames.fromclasstogame.core.entity.GameObject;
 import fr.snapgames.fromclasstogame.core.exceptions.io.UnknownResource;
+import fr.snapgames.fromclasstogame.core.gfx.renderer.RenderHelper;
 import fr.snapgames.fromclasstogame.core.io.ResourceManager;
 import fr.snapgames.fromclasstogame.demo.entity.InventoryObject;
